:arrows_clockwise: Earlier Node.js versions are supported
New command-line option is added, and

- :grinning: works on Node.js v5.10.0 or later
- :grinning: works on Node.js v4.5.0 or later, if major version is 4
- :disappointed: doesn't work on Node.js earlier than v5.10.0
- :disappointed: doesn't work on Node.js earlier than v4.5.0

## NOTE

To configure `git` to use `emojify` as pager; for example, on :penguin: linux,

```shell
git config --global core.pager \
	"sh -c 'emojify --avoid-includes --avoid-subarray | less -R'"
```

If this option is specified, `emojify` avoid writing data,
instanciated by Buffer.prototype.subarray, to standard output.

## NOTICE

This option is suitable for Node.js earlier than v6.0.0,
however, it **SHOULD NOT** be used for v6.0.0 or later.

Signed-off-by: kei-g <km.8k6ce+github@gmail.com>