:arrows_clockwise: Earlier Node.js versions are supported
New command-line option is added, and

- :grinning: works on Node.js v4.0.0 or later
- :disappointed: doesn't work on Node.js earlier than v4.0.0

## NOTE

To configure `git` to use `emojify` as pager; for example, on :penguin: linux,

```shell
git config --global core.pager \
	"sh -c 'emojify --avoid-buffer-alloc --avoid-includes --avoid-subarray | less -R'"
```

If this option is specified, `emojify` avoid allocating Buffer by
Buffer.alloc, and uses new Buffer instead.

## NOTICE

This option is suitable for Node.js earlier than v5.10.0 or v4.5.0,
however, it **SHOULD NOT** be used for v4.5.0 or later, v5.10.0 or later.

Signed-off-by: kei-g <km.8k6ce+github@gmail.com>